calcChange moved to BoxSelector
